# 2 - Indroducao ao React/EventPlus/eventplusapi/apiweb.eventplus.manha/Controllers/EventoController.cs
using apiweb.eventplus.manha.Domains;
using apiweb.eventplus.manha.Interfaces;
using apiweb.eventplus.manha.Repositories;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;

namespace apiweb.eventplus.manha.Controllers
{
    [Route("api/[controller]")]
    [ApiController]
    [Produces("application/json")]
    public class EventoController : ControllerBase
    {
        IEventoRepository _eventoRepository;

        public EventoController()
        {
            _eventoRepository = new EventoRepository();
        }

        [HttpGet("ListarProximos")]
        public IActionResult ListarProximos()
        {
            try
            {
                return Ok(_eventoRepository.ListarProximos());
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
            }
        }

        [HttpGet("ListarAnteriores")]
        public IActionResult ListarAnteriores()
        {
            try
            {
                return Ok(_eventoRepository.ListarAnteriores());
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
            }
        }

        [HttpGet]
        public IActionResult Get()
        {
            try
            {
                return Ok(_eventoRepository.Listar());
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
                throw;
            }
        }

        [HttpPost]
        public IActionResult Post(Evento evento)
        {
            try
            {
                if (evento != null)
                {
                    _eventoRepository.Cadastrar(evento);
                    return Ok("Evento cadastrado!");
                }

                return Ok("Evento digitado incorretamente");
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
                throw;
            }
        }

        [HttpDelete("{id}")]
        public IActionResult Delete(Guid id)
        {
            try
            {
                _eventoRepository.Delete(id);
                return Ok("Evento deletado!");
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
            }
        }

        [HttpGet("{id}")]
        public IActionResult GetWithId(Guid id)
        {
            try
            {
                return Ok(_eventoRepository.BuscarPorId(id));
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
            }
        }

        [HttpPut("{id}")]
        public IActionResult Put(Guid id, Evento evento)
        {
            try
            {
                _eventoRepository.Atualizar(id, evento);
                return Ok("Evento atualizado!");
            }
            catch (Exception e)
            {
                return BadRequest(e.Message);
            }
        }


    }
}